Merge datasets.
This is an example on how to merge data fetched from regParams "curr" key and sysParams.

This should be considered unsafe merge as it overrides existing keys and should only be used for learning purposes.

For proper implementation - both endpoints should be merged under different keys and accessed via separate entity sub-classes.